refactored access and range
